Retire ag-grid
