+++ package.json
@@ -59,8 +59,6 @@
     "@tiptap/starter-kit": "^3.22.5",
     "@typescript-eslint/eslint-plugin": "^8.58.0",
     "@xyflow/react": "^12.10.2",
-    "ag-grid-community": "^35.2.1",
-    "ag-grid-react": "^35.2.1",
     "better-auth": "^1.6.3",
     "bootstrap": "^5.3.8",
     "chalk": "^5.6.2",
