+++ src/components/model-editor/metric-viewer/MetricDataViewer.tsx
@@ -12,9 +12,16 @@ import {
   Typography,
 } from '@mui/material';
 
-import type { CellClassParams, ColDef } from 'ag-grid-community';
+import {
+  DataEditor,
+  type DataEditorProps,
+  type GridCell,
+  GridCellKind,
+  type GridColumn,
+  type Item,
+} from '@glideapps/glide-data-grid';
+import '@glideapps/glide-data-grid/dist/index.css';
 
-import AgGridReact from '../GridEditor';
 import {
   type DimensionalMetric,
   type MetricCategory,
@@ -29,7 +36,21 @@ type MetricDataViewerProps = {
   fillHeight?: boolean;
 };
 
-const FORECAST_CELL_CLASS = 'metric-forecast-cell';
+type SortState = { colId: string; dir: 'asc' | 'desc' } | null;
+
+type ColumnSpec = {
+  col: GridColumn & { id: string };
+  sortable?: boolean;
+  getCell: (rowIndex: number) => GridCell;
+  getSortValue?: (rowIndex: number) => number | string | null;
+};
+
+const FORECAST_OVERRIDE = { bgCell: '#f1f8fe' } as const;
+
+function formatNumber(value: number | null | undefined, maxDigits: number): string {
+  if (value == null) return '';
+  return value.toLocaleString(undefined, { maximumFractionDigits: maxDigits });
+}
 
 export default function MetricDataViewer({
   metric,
@@ -39,8 +60,9 @@ export default function MetricDataViewer({
   const hasDimensions = metric.dimensions.length > 0;
   const [viewMode, setViewMode] = useState<ViewMode>(hasDimensions ? 'pivot' : 'flat');
   const [pivotDimId, setPivotDimId] = useState<string>(() => metric.dimensions[0]?.id ?? '');
-
   const [filters, setFilters] = useState<Record<string, string[]>>({});
+  const [sort, setSort] = useState<SortState>(null);
+  const [colWidths, setColWidths] = useState<Record<string, number>>({});
 
   const handleFilterChange = useCallback((dimId: string, catIds: string[]) => {
     setFilters((prev) => ({ ...prev, [dimId]: catIds }));
@@ -53,29 +75,49 @@ export default function MetricDataViewer({
     return metric.toPivotRows(pivotDimId);
   }, [metric, viewMode, pivotDimId]);
 
-  const pivotColumnDefs = useMemo<ColDef[]>(() => {
+  const pivotSpecs = useMemo<ColumnSpec[]>(() => {
     if (!pivotData) return [];
-    const yearCol: ColDef = {
-      field: 'year',
-      headerName: 'Year',
-      pinned: 'left',
-      width: 80,
-      sortable: true,
+    const rows = pivotData.rows as Array<Record<string, unknown>>;
+    const isForecastRow = (rowIndex: number): boolean => {
+      if (forecastFrom == null) return false;
+      const year = rows[rowIndex]?.year as number | undefined;
+      return year != null && year >= forecastFrom;
     };
-    const catCols: ColDef[] = pivotData.columns.map((cat: MetricCategory) => ({
-      field: cat.id,
-      headerName: cat.label,
-      type: 'numericColumn',
-      width: 120,
-      valueFormatter: (params: { value: number | null }) =>
-        params.value != null
-          ? params.value.toLocaleString(undefined, { maximumFractionDigits: 2 })
-          : '',
-      cellClass: (params: CellClassParams) =>
-        forecastFrom != null && params.data?.year >= forecastFrom ? FORECAST_CELL_CLASS : '',
+    const yearSpec: ColumnSpec = {
+      col: { id: 'year', title: 'Year', width: colWidths.year ?? 80 },
+      sortable: true,
+      getCell: (rowIndex) => {
+        const value = rows[rowIndex]?.year as number | undefined;
+        return {
+          kind: GridCellKind.Number,
+          data: value,
+          displayData: value != null ? String(value) : '',
+          allowOverlay: false,
+          readonly: true,
+          contentAlign: 'right',
+          themeOverride: isForecastRow(rowIndex) ? FORECAST_OVERRIDE : undefined,
+        };
+      },
+      getSortValue: (rowIndex) => (rows[rowIndex]?.year as number | undefined) ?? null,
+    };
+    const catSpecs: ColumnSpec[] = pivotData.columns.map((cat: MetricCategory) => ({
+      col: { id: cat.id, title: cat.label, width: colWidths[cat.id] ?? 120 },
+      getCell: (rowIndex) => {
+        const value = rows[rowIndex]?.[cat.id] as number | null | undefined;
+        return {
+          kind: GridCellKind.Number,
+          data: value ?? undefined,
+          displayData: formatNumber(value, 2),
+          allowOverlay: false,
+          readonly: true,
+          contentAlign: 'right',
+          themeOverride: isForecastRow(rowIndex) ? FORECAST_OVERRIDE : undefined,
+        };
+      },
+      getSortValue: (rowIndex) => (rows[rowIndex]?.[cat.id] as number | null | undefined) ?? null,
     }));
-    return [yearCol, ...catCols];
-  }, [pivotData, forecastFrom]);
+    return [yearSpec, ...catSpecs];
+  }, [pivotData, forecastFrom, colWidths]);
 
   const filteredMetric = useMemo(() => {
     if (viewMode !== 'flat') return metric;
@@ -90,40 +132,141 @@ export default function MetricDataViewer({
 
   const flatRows = useMemo(() => {
     if (viewMode !== 'flat') return [];
-    return filteredMetric.toRows();
+    return filteredMetric.toRows() as Array<Record<string, unknown>>;
   }, [filteredMetric, viewMode]);
 
-  const flatColumnDefs = useMemo<ColDef[]>(() => {
+  const flatSpecs = useMemo<ColumnSpec[]>(() => {
     if (viewMode !== 'flat') return [];
-    const dimCols: ColDef[] = filteredMetric.dimensions.map((dim: MetricDimension) => {
+    const isForecastRow = (rowIndex: number): boolean => {
+      if (forecastFrom == null) return false;
+      const year = flatRows[rowIndex]?.year as number | undefined;
+      return year != null && year >= forecastFrom;
+    };
+    const dimSpecs: ColumnSpec[] = filteredMetric.dimensions.map((dim: MetricDimension) => {
       const catLabelMap = new Map(dim.categories.map((c) => [c.id, c.label]));
       return {
-        field: dim.id,
-        headerName: dim.label,
-        width: 140,
+        col: { id: dim.id, title: dim.label, width: colWidths[dim.id] ?? 140 },
         sortable: true,
-        valueFormatter: (params: { value: string }) =>
-          catLabelMap.get(params.value) ?? params.value,
+        getCell: (rowIndex) => {
+          const raw = flatRows[rowIndex]?.[dim.id] as string | undefined;
+          const label = (raw && catLabelMap.get(raw)) ?? raw ?? '';
+          return {
+            kind: GridCellKind.Text,
+            data: label,
+            displayData: label,
+            allowOverlay: false,
+            readonly: true,
+            themeOverride: isForecastRow(rowIndex) ? FORECAST_OVERRIDE : undefined,
+          };
+        },
+        getSortValue: (rowIndex) => {
+          const raw = flatRows[rowIndex]?.[dim.id] as string | undefined;
+          return (raw && catLabelMap.get(raw)) ?? raw ?? null;
+        },
       };
     });
-    const yearCol: ColDef = { field: 'year', headerName: 'Year', width: 80, sortable: true };
-    const valCol: ColDef = {
-      field: 'value',
-      headerName: `Value (${metric.unit.short})`,
-      type: 'numericColumn',
-      width: 130,
+    const yearSpec: ColumnSpec = {
+      col: { id: 'year', title: 'Year', width: colWidths.year ?? 80 },
       sortable: true,
-      valueFormatter: (params: { value: number | null }) =>
-        params.value != null
-          ? params.value.toLocaleString(undefined, { maximumFractionDigits: 4 })
-          : '',
-      cellClass: (params: CellClassParams) =>
-        forecastFrom != null && params.data?.year >= forecastFrom ? FORECAST_CELL_CLASS : '',
+      getCell: (rowIndex) => {
+        const value = flatRows[rowIndex]?.year as number | undefined;
+        return {
+          kind: GridCellKind.Number,
+          data: value,
+          displayData: value != null ? String(value) : '',
+          allowOverlay: false,
+          readonly: true,
+          contentAlign: 'right',
+          themeOverride: isForecastRow(rowIndex) ? FORECAST_OVERRIDE : undefined,
+        };
+      },
+      getSortValue: (rowIndex) => (flatRows[rowIndex]?.year as number | undefined) ?? null,
     };
-    return [...dimCols, yearCol, valCol];
-  }, [filteredMetric, viewMode, metric.unit.short, forecastFrom]);
+    const valSpec: ColumnSpec = {
+      col: {
+        id: 'value',
+        title: `Value (${metric.unit.short})`,
+        width: colWidths.value ?? 130,
+      },
+      sortable: true,
+      getCell: (rowIndex) => {
+        const value = flatRows[rowIndex]?.value as number | null | undefined;
+        return {
+          kind: GridCellKind.Number,
+          data: value ?? undefined,
+          displayData: formatNumber(value, 4),
+          allowOverlay: false,
+          readonly: true,
+          contentAlign: 'right',
+          themeOverride: isForecastRow(rowIndex) ? FORECAST_OVERRIDE : undefined,
+        };
+      },
+      getSortValue: (rowIndex) => (flatRows[rowIndex]?.value as number | null | undefined) ?? null,
+    };
+    return [...dimSpecs, yearSpec, valSpec];
+  }, [filteredMetric, viewMode, flatRows, metric.unit.short, forecastFrom, colWidths]);
+
+  const isPivot = viewMode === 'pivot' && !!pivotData;
+  const specs = isPivot ? pivotSpecs : flatSpecs;
+  const rawRowCount = isPivot ? (pivotData?.rows.length ?? 0) : flatRows.length;
+
+  const sortedIndex = useMemo<number[] | null>(() => {
+    if (!sort) return null;
+    const spec = specs.find((s) => s.col.id === sort.colId);
+    if (!spec?.getSortValue) return null;
+    const indices = Array.from({ length: rawRowCount }, (_, i) => i);
+    const factor = sort.dir === 'asc' ? 1 : -1;
+    indices.sort((a, b) => {
+      const va = spec.getSortValue!(a);
+      const vb = spec.getSortValue!(b);
+      if (va == null && vb == null) return 0;
+      if (va == null) return 1;
+      if (vb == null) return -1;
+      if (typeof va === 'number' && typeof vb === 'number') return (va - vb) * factor;
+      return String(va).localeCompare(String(vb)) * factor;
+    });
+    return indices;
+  }, [sort, specs, rawRowCount]);
+
+  const getCellContent = useCallback<DataEditorProps['getCellContent']>(
+    (cell: Item) => {
+      const [col, row] = cell;
+      const realRow = sortedIndex ? (sortedIndex[row] ?? row) : row;
+      const spec = specs[col];
+      if (!spec) {
+        return { kind: GridCellKind.Loading, allowOverlay: false } as GridCell;
+      }
+      return spec.getCell(realRow);
+    },
+    [specs, sortedIndex]
+  );
+
+  const onHeaderClicked = useCallback<NonNullable<DataEditorProps['onHeaderClicked']>>(
+    (colIndex) => {
+      const spec = specs[colIndex];
+      if (!spec?.sortable) return;
+      const id = spec.col.id;
+      setSort((prev) => {
+        if (!prev || prev.colId !== id) return { colId: id, dir: 'asc' };
+        if (prev.dir === 'asc') return { colId: id, dir: 'desc' };
+        return null;
+      });
+    },
+    [specs]
+  );
+
+  const onColumnResize = useCallback<NonNullable<DataEditorProps['onColumnResize']>>(
+    (column, newSize) => {
+      if (!column.id) return;
+      setColWidths((prev) => ({ ...prev, [column.id as string]: newSize }));
+    },
+    []
+  );
+
+  const columns = useMemo<GridColumn[]>(() => specs.map((s) => s.col), [specs]);
 
   const gridHeight = compact ? 300 : 500;
+  const freezeColumns = isPivot ? 1 : 0;
 
   return (
     <Box
@@ -154,7 +297,7 @@ export default function MetricDataViewer({
           <ToggleButtonGroup
             value={viewMode}
             exclusive
-            onChange={(_, v) => {
+            onChange={(_, v: ViewMode | null) => {
               if (v) setViewMode(v);
             }}
             size="small"
@@ -208,32 +351,25 @@ export default function MetricDataViewer({
       )}
 
       <Box
-        className="ag-theme-alpine"
         sx={{
           ...(fillHeight ? { flex: 1, minHeight: 0 } : { height: gridHeight }),
           width: '100%',
-          [`& .${FORECAST_CELL_CLASS}`]: {
-            backgroundColor: 'rgba(33, 150, 243, 0.06)',
-          },
         }}
       >
-        {viewMode === 'pivot' && pivotData && (
-          <AgGridReact
-            columnDefs={pivotColumnDefs}
-            rowData={pivotData.rows}
-            defaultColDef={{ resizable: true, suppressMovable: true }}
-            suppressColumnVirtualisation={compact && !fillHeight}
-            domLayout={compact && !fillHeight ? 'autoHeight' : undefined}
-          />
-        )}
-        {viewMode === 'flat' && (
-          <AgGridReact
-            columnDefs={flatColumnDefs}
-            rowData={flatRows}
-            defaultColDef={{ resizable: true, suppressMovable: true }}
-            domLayout={compact && !fillHeight ? 'autoHeight' : undefined}
-          />
-        )}
+        <DataEditor
+          columns={columns}
+          rows={rawRowCount}
+          getCellContent={getCellContent}
+          onHeaderClicked={onHeaderClicked}
+          onColumnResize={onColumnResize}
+          freezeColumns={freezeColumns}
+          smoothScrollX
+          smoothScrollY
+          width="100%"
+          height="100%"
+          rowHeight={32}
+          headerHeight={36}
+        />
       </Box>
     </Box>
   );
